fix: author element 중복 제거
