+++ src/main/java/com/ssafy/ssafsound/domain/comment/dto/AuthorElement.java
@@ -3,6 +3,8 @@ package com.ssafy.ssafsound.domain.comment.dto;
 import com.ssafy.ssafsound.domain.comment.domain.Comment;
 import com.ssafy.ssafsound.domain.member.domain.Member;
 import com.ssafy.ssafsound.domain.member.dto.SSAFYInfo;
+import com.ssafy.ssafsound.domain.post.domain.Post;
+
 import lombok.Getter;
 
 @Getter
@@ -33,6 +35,17 @@ public class AuthorElement {
         this.ssafyMember = member.getSsafyMember();
         this.isMajor = member.getMajor();
         this.ssafyInfo = SSAFYInfo.from(member);
+    }
+
+    public static AuthorElement from(Post post) {
+        Member member = post.getMember();
+        if (post.getAnonymity())
+            return new AuthorElement("익명");
+        return new AuthorElement(member);
+    }
+
+    private AuthorElement(String nickname) {
+        this.nickname = nickname;
     }
 
     private String setNickName(Comment comment) {
